requirements.txt
asgiref==3.5.0
click==8.1.2
clt==1.1
colorthief==0.2.1
cycler==0.11.0
Django==4.0.4
dominate==2.6.0
Flask==2.1.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.0.1
fonttools==4.32.0
image==1.5.33
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
kmeans==1.0.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
visitor==0.1.3
webcolors==1.11.1
Werkzeug==2.1.1
WTForms==3.0.1
